+++ bot/src/main/kotlin/org/epilink/bot/rulebook/QuickHttp.kt
@@ -9,24 +9,24 @@
 package org.epilink.bot.rulebook
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.apache.Apache
-import io.ktor.client.features.auth.Auth
-import io.ktor.client.features.auth.providers.basic
-import io.ktor.client.request.get
-import io.ktor.client.request.header
-import io.ktor.http.ContentType
-import io.ktor.http.HttpHeaders
+import io.ktor.client.*
+import io.ktor.client.engine.apache.*
+import io.ktor.client.features.auth.*
+import io.ktor.client.features.auth.providers.*
+import io.ktor.client.request.*
+import io.ktor.http.*
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.slf4j.LoggerFactory
+import kotlin.reflect.KClass
 
 /*
  * Utility functions for use within a Rulebook
  */
 
 private val logger = LoggerFactory.getLogger("epilink.rulebooks.helpers")
+
+// TODO test these functions
 
 /**
  * Perform a HTTP GET request to the given URL, expecting a JSON reply, and return this JSON reply as a Map from String
@@ -37,14 +37,62 @@ private val logger = LoggerFactory.getLogger("epilink.rulebooks.helpers")
  * @param bearer The bearer to use in the Authorization header, or null to not do that
  * @param eagerAuthentication If the authentication information should be sent directly or only after the server sends
  * back an unauthorized error
+ * @param headers Custom headers that will be added to the request
  */
 @Suppress("unused")
 suspend fun httpGetJson(
     url: String,
     basicAuth: Pair<String, String>? = null,
     bearer: String? = null,
-    eagerAuthentication: Boolean = false
-): Map<String, Any?> {
+    eagerAuthentication: Boolean = true,
+    headers: List<Pair<String, String>> = listOf()
+): Map<String, Any?> =
+    httpGetJsonTyped(url, basicAuth, bearer, eagerAuthentication, headers)
+
+/**
+ * Perform a HTTP GET request to the given URL, expecting a JSON reply, and return this JSON reply as an object of type
+ * `T`
+ *
+ * @param url The URL for the GET request
+ * @param basicAuth The Basic authentication credentials to use, or null to not use credentials at all
+ * @param bearer The bearer to use in the Authorization header, or null to not do that
+ * @param eagerAuthentication If the authentication information should be sent directly or only after the server sends
+ * back an unauthorized error
+ * @param headers Custom headers that will be added to the request
+ * @param T The type of the value that is returned by the JSON API. Can be, for example, `List<*>`, `String`, or a
+ * custom data class.
+ */
+suspend inline fun <reified T : Any> httpGetJsonTyped(
+    url: String,
+    basicAuth: Pair<String, String>? = null,
+    bearer: String? = null,
+    eagerAuthentication: Boolean = true,
+    headers: List<Pair<String, String>> = listOf()
+): T =
+    httpGetJsonClass(url, T::class, basicAuth, bearer, eagerAuthentication, headers)
+
+/**
+ * Perform a HTTP GET request to the given URL, expecting a JSON reply, and return this JSON reply as an instance of
+ * the class given in the parameters.
+ *
+ * @param url The URL for the GET request
+ * @param returnType The class representing the type of what the URL is supposed to return.
+ * @param basicAuth The Basic authentication credentials to use, or null to not use credentials at all
+ * @param bearer The bearer to use in the Authorization header, or null to not do that
+ * @param eagerAuthentication If the authentication information should be sent directly or only after the server sends
+ * back an unauthorized error
+ * @param headers Custom headers that will be added to the request
+ * @param T The type of the value that is returned by the JSON API. Can be, for example, `List<*>`, `String`, or a
+ * custom data class.
+ */
+suspend fun <T : Any> httpGetJsonClass(
+    url: String,
+    returnType: KClass<T>,
+    basicAuth: Pair<String, String>? = null,
+    bearer: String? = null,
+    eagerAuthentication: Boolean = true,
+    headers: List<Pair<String, String>> = listOf()
+): T {
     val client = HttpClient(Apache) {
         install(Auth) {
             if (basicAuth != null) {
@@ -59,6 +107,7 @@ suspend fun httpGetJson(
     val response = runCatching {
         client.get<String>(url) {
             header(HttpHeaders.Accept, ContentType.Application.Json)
+            headers.forEach { (name, value) -> header(name, value) }
             if (bearer != null)
                 header(HttpHeaders.Authorization, "Bearer $bearer")
         }
@@ -67,7 +116,7 @@ suspend fun httpGetJson(
         throw RuleException("Encountered an error on httpGetJson call", it)
     }
 
-    return withContext(Dispatchers.Default) { ObjectMapper().readValue(response) }
+    return withContext(Dispatchers.Default) { ObjectMapper().readValue(response, returnType.java) }
 }
 
 /**
@@ -88,14 +137,23 @@ fun Map<*, *>.getString(key: String) = this[key] as? String ?: error("Invalid fo
 /**
  * Get a value at an index, where the value is expected to be a map
  */
-fun List<*>.getMap(index: Int) = this[index] as? Map<*, *> ?: error("Invalid format, value at index $index is not a map")
+fun List<*>.getMap(index: Int) =
+    this[index] as? Map<*, *> ?: error("Invalid format, value at index $index is not a map")
 
 /**
  * Get the value at an index, where the value is expected to be a list
  */
-fun List<*>.getList(index: Int) = this[index] as? List<*> ?: error("Invalid format, value at index $index is not a list")
+fun List<*>.getList(index: Int) =
+    this[index] as? List<*> ?: error("Invalid format, value at index $index is not a list")
 
 /**
  * Get the value at an index, where the value is expected to be a string
  */
-fun List<*>.getString(index: Int) = this[index] as? String ?: error("Invalid format, value at index $index is not a string")
+fun List<*>.getString(index: Int) =
+    this[index] as? String ?: error("Invalid format, value at index $index is not a string")
+
+/**
+ * Equivalent to [map], but casts each element to [T] each. Returns a list of elements of type [R]
+ */
+inline fun <reified T, R> Iterable<*>.mapAs(block: (T) -> R): List<R> =
+    map { block(it as? T ?: error("Invalid format, value $it is not of type T (${T::class.qualifiedName}")) }